refactor: Move more boolean code to polars_compute, reusing kernels
