+++ Cargo.toml
@@ -152,8 +152,6 @@ features = [
   "compute_aggregate",
   "compute_arithmetics",
   "compute_bitwise",
-  "compute_boolean",
-  "compute_boolean_kleene",
   "compute_comparison",
 ]
