+++ crates/polars-arrow/Cargo.toml
@@ -104,16 +104,12 @@ compute_aggregate = []
 compute_arithmetics_decimal = []
 compute_arithmetics = ["compute_arithmetics_decimal"]
 compute_bitwise = []
-compute_boolean = []
-compute_boolean_kleene = []
-compute_comparison = ["compute_boolean"]
+compute_comparison = []
 compute_temporal = []
 compute = [
   "compute_aggregate",
   "compute_arithmetics",
   "compute_bitwise",
-  "compute_boolean",
-  "compute_boolean_kleene",
   "compute_comparison",
   "compute_temporal",
 ]
