+++ crates/polars-arrow/src/compute/mod.rs
@@ -19,12 +19,6 @@ pub mod arity_assign;
 #[cfg(feature = "compute_bitwise")]
 #[cfg_attr(docsrs, doc(cfg(feature = "compute_bitwise")))]
 pub mod bitwise;
-#[cfg(feature = "compute_boolean")]
-#[cfg_attr(docsrs, doc(cfg(feature = "compute_boolean")))]
-pub mod boolean;
-#[cfg(feature = "compute_boolean_kleene")]
-#[cfg_attr(docsrs, doc(cfg(feature = "compute_boolean_kleene")))]
-pub mod boolean_kleene;
 pub mod concatenate;
 #[cfg(feature = "dtype-decimal")]
 pub mod decimal;
