+++ crates/polars-ops/src/chunked_array/array/any_all.rs
@@ -4,7 +4,7 @@ use arrow::legacy::utils::CustomIterTools;
 
 use super::*;
 
-fn array_all_any<F>(arr: &FixedSizeListArray, op: F, is_all: bool) -> PolarsResult<BooleanArray>
+fn array_all_any<F>(arr: &FixedSizeListArray, op: F, fast_value: bool) -> PolarsResult<BooleanArray>
 where
     F: Fn(&BooleanArray) -> bool,
 {
@@ -15,13 +15,10 @@ where
     let values = values.as_any().downcast_ref::<BooleanArray>().unwrap();
     let validity = arr.validity().cloned();
 
-    // Fast path where all values set (all is free).
-    if is_all {
-        let all_set = arrow::compute::boolean::all(values);
-        if all_set {
-            let bits = Bitmap::new_with_value(true, arr.len());
-            return Ok(BooleanArray::from_data_default(bits, None).with_validity(validity));
-        }
+    // Fast path where all/none values set.
+    if op(values) == fast_value {
+        let bits = Bitmap::new_with_value(fast_value, arr.len());
+        return Ok(BooleanArray::from_data_default(bits, None).with_validity(validity));
     }
 
     let len = arr.size();
@@ -39,15 +36,23 @@ where
 }
 
 pub(super) fn array_all(ca: &ArrayChunked) -> PolarsResult<Series> {
-    let chunks = ca
-        .downcast_iter()
-        .map(|arr| array_all_any(arr, arrow::compute::boolean::all, true));
+    let chunks = ca.downcast_iter().map(|arr| {
+        array_all_any(
+            arr,
+            |a| polars_compute::boolean::all(a).unwrap_or(true),
+            true,
+        )
+    });
     Ok(BooleanChunked::try_from_chunk_iter(ca.name().clone(), chunks)?.into_series())
 }
 
 pub(super) fn array_any(ca: &ArrayChunked) -> PolarsResult<Series> {
-    let chunks = ca
-        .downcast_iter()
-        .map(|arr| array_all_any(arr, arrow::compute::boolean::any, false));
+    let chunks = ca.downcast_iter().map(|arr| {
+        array_all_any(
+            arr,
+            |a| polars_compute::boolean::any(a).unwrap_or(false),
+            false,
+        )
+    });
     Ok(BooleanChunked::try_from_chunk_iter(ca.name().clone(), chunks)?.into_series())
 }
