+++ crates/polars/tests/it/arrow/compute/mod.rs
@@ -1,6 +1,4 @@
 mod aggregate;
 mod bitwise;
-mod boolean;
-mod boolean_kleene;
 
 mod arity_assign;
